+++ dracut/modules.d/90wwinit/load-wwinit.sh
@@ -3,18 +3,25 @@
 info "Mounting tmpfs at $NEWROOT"
 mount -t tmpfs -o mpol=interleave ${wwinit_tmpfs_size_option} tmpfs "$NEWROOT"
 
-for archive in "${wwinit_image}" "${wwinit_system}" "${wwinit_runtime}"
+for stage in "image" "system" "runtime"
 do
-    if [ -n "${archive}" ]
+    info "Loading stage: ${stage}"
+    # Load runtime overlay from a static privledged port.
+    # Others use default settings.
+    localport=""
+    if [[ "${stage}" == "runtime" ]]
     then
-        info "Loading ${archive}"
-        # Load runtime overlay from a static privledged port.
-        # Others use default settings.
-        localport=""
-        if [[ "${archive}" == "${wwinit_runtime}" ]]
-        then
-            localport="--local-port 1-1023"
-        fi
-        (curl --retry 60 --retry-delay 1 --silent ${localport} -L "${archive}" | gzip -d | cpio -im --directory="${NEWROOT}") || die "Unable to load ${archive}"
+        localport="--local-port 1-1023"
     fi
+    (
+        curl --location --silent --get ${localport} \
+            --retry 60 --retry-delay 1 \
+            --data-urlencode "assetkey=${wwinit_assetkey}" \
+            --data-urlencode "uuid=${wwinit_uuid}" \
+            --data-urlencode "stage=${stage}" \
+            --data-urlencode "compress=gz" \
+            "${wwinit_uri}" \
+        | gzip -d \
+        | cpio -im --directory="${NEWROOT}"
+    ) || die "Unable to load stage: ${stage}"
 done